+++ src/main/java/edu/eci/pdsw/samples/entities/Estudiante.java
@@ -14,7 +14,7 @@ public class Estudiante extends Persona{
     private int codigoEstudiante,semestrePonderado;
     private String carrera;
     
-    public Estudiante(int documentoID, int telefono1, int telefono2, String tipoDocumentoID, String nombre, String direccion, String correo, String genero,int codigoEstudiante,int semestrePonderado, String carrera) {
+    public Estudiante(int documentoID, long telefono1, long telefono2, String tipoDocumentoID, String nombre, String direccion, String correo, String genero,int codigoEstudiante,int semestrePonderado, String carrera) {
         super(documentoID,telefono1,telefono2,tipoDocumentoID,nombre,direccion,correo,genero);
         this.codigoEstudiante = codigoEstudiante;
         this.semestrePonderado = semestrePonderado;
